+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2020.3.25f1
-m_EditorVersionWithRevision: 2020.3.25f1 (9b9180224418)
+m_EditorVersion: 2020.3.24f1
+m_EditorVersionWithRevision: 2020.3.24f1 (79c78de19888)
